requirements.txt
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-kms==2.15.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
requests==2.28.2
rsa==4.9
six==1.16.0
urllib3==1.26.15
